2002-09-16  Guillaume Pelat  <endymion@linux-secure.com>
        * sensor-adduser/plaintext.c(setup_plaintext):
        * src/common.c(prelude_read_multiline):
        * src/config-engine.c(cmp_entry):
        (create_new_line):
        * src/prelude-client.c(handle_plaintext_connection):
        * src/prelude-getopt.c(check_option_optarg):
        (check_option_reqarg):
        (lookup_variable_if_needed):
        * src/ssl-gencrypto.c(get_full_hostname):
        Fix potential signed/unsigned integer problems.


git-svn-id: file:///home/yoann/SVN/trunk/libprelude@1612 09c5ec92-17d4-0310-903a-819935f44dba

+++ sensor-adduser/plaintext.c
@@ -56,7 +56,7 @@ static int setup_plaintext(prelude_io_t *fd, const char *oneshot,
 {
         ssize_t ret;
         prelude_msg_t *msg;
-        int len, ulen, plen, olen;
+        size_t len, ulen, plen, olen;
 
         ulen = strlen(user) + 1;
         plen = strlen(pass) + 1;
